fix: merge FB layouts upgrade into a single process
