+++ 5.38.0/updateDefaultFormLayout.ts
@@ -9,17 +9,37 @@ const ncp = util.promisify(ncpBase.ncp);
 export const updateDefaultFormLayout = createProcessor(async params => {
     const { context } = params;
 
-    context.log.info(
-        "Migrating form layouts located in the %s folder.",
-        "apps/theme/layouts/forms"
+    context.log.info("Upgrading form layouts located in %s folder...", "apps/theme/layouts/forms");
+
+    log.info("Backing up %s folder...", "apps/theme");
+
+    const fromBackup = path.join(params.context.project.root, "apps", "theme", "layouts", "forms");
+    const toBackup = path.join(
+        params.context.project.root,
+        "apps",
+        "theme",
+        "layouts",
+        "_forms_backup"
     );
 
-    const from = path.join(__dirname, "updateDefaultFormLayout");
-    const to = path.join(context.project.root, "apps", "theme", "layouts", "forms");
-
-    if (!fs.existsSync(to)) {
+    if (fs.existsSync(toBackup)) {
         log.warning(
-            "Cannot perform migration because the %s folder doesn't exist. This could be because the %s wasn't performed. To learn more, visit https://www.webiny.com/docs/release-notes/5.34.0/page-builder-pe-rendering-engine-migration.",
+            "%s folder already exists, cannot create backup of the %s folder. Skipping form layouts upgrade...",
+            toBackup,
+            "theme"
+        );
+        return;
+    }
+
+    fs.mkdirSync(toBackup);
+    await ncp(fromBackup, toBackup);
+
+    const newCodeFrom = path.join(__dirname, "updateDefaultFormLayout");
+    const newCodeTo = path.join(context.project.root, "apps", "theme", "layouts", "forms");
+
+    if (!fs.existsSync(newCodeTo)) {
+        log.warning(
+            "Cannot perform upgrade because the %s folder doesn't exist. This could be because the %s wasn't performed. To learn more, visit https://www.webiny.com/docs/release-notes/5.34.0/page-builder-pe-rendering-engine-migration.",
             "apps/theme/layouts/forms",
             "Page Builder - Page Rendering Engine Migration"
         );
@@ -31,5 +51,7 @@ export const updateDefaultFormLayout = createProcessor(async params => {
         "Generating latest code in the form layouts folder (%s).",
         "apps/theme/layouts/forms"
     );
-    await ncp(from, to);
+    await ncp(newCodeFrom, newCodeTo);
+
+    log.success("Form layouts upgrade completed successfully.");
 });
